Module Private.Commit

Private commit store.

Commit Store

type -'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key = Hash.t

The type for keys.

type value

The type for raw values.

val mem : [> Irmin.Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin.Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin.Perms.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin.Perms.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

val merge : [> Irmin.Perms.read_write ] t -> info:Irmin.Info.f -> key option Irmin.Merge.t

merge is the 3-way merge function for commit keys.

module Key : Irmin.Hash.TYPED with type t = key and type value = value

Key provides base functions for commit keys.

module Val : sig ... end

Val provides functions for commit values.

module Node : sig ... end

Node is the underlying node store.