Module Tree.Contents

Operations on lazy tree contents.

type t

The type of lazy tree contents.

val hash : t -> hash

hash t is the hash of the contents value returned when t is forced successfully.

val force : t -> contents or_error Lwt.t

force t forces evaluation of the lazy content value t, or returns an error if no such value exists in the underlying repository.

val clear : t -> unit

clear t clears t's cache.