Module Private.Commit

Commit Store

include Irmin__.S.CONTENT_ADDRESSABLE_STORE

type 'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> `Read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> `Read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> `Write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> `Write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

val merge : [ `Read | `Write ] t -> info:Irmin.Info.f -> key option Irmin.Merge.t

merge is the 3-way merge function for commit keys.

module Key : Irmin__.S.TYPED_HASH with type t = key and type value = value

Key provides base functions for commit keys.

module Val : Irmin__.S.COMMIT with type t = value and type hash = key and type hash = Node.key

Val provides functions for commit values.

module Node : Irmin__.S.NODE_STORE with type key = Val.hash

Node is the underlying node store.